refactor(cpuhealth): give the vocabulary, the decision and its helpers their own files
decide.go held three unrelated things in one 423-line file: the vocabulary a
verdict is written in (State, Attribution, CauseKind, Unit, Cause, Verdict,
saturationArm), the decision itself, and the two helpers plus the Signals fact
bag it fills. Each now has a file — verdict.go, decide.go, attribute.go,
signals.go — so reaching Decide no longer means scrolling past 68 lines of type
declarations, and the 38-field Signals struct no longer sits behind it.

Pure move: every declaration and its doc comment travel unedited, nothing is
renamed, and no test changes. The four files reassemble byte-for-byte into the
original decide.go, and `go doc -all -u ./pkg/cpuhealth` hashes unchanged.

cpuReserveCores moves to verdict.go rather than staying with Decide because
table.go reads it too.
