+++ umh-core/pkg/cpuhealth/verdict.go
@@ -1,0 +1,84 @@
+// Copyright 2025 UMH Systems GmbH
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cpuhealth
+
+// State is the customer-visible health state of the CPU verdict. Two values and
+// no third: the dead zone is an annotation on a healthy verdict, never a state.
+type State string
+
+const (
+	StateHealthy  State = "healthy"
+	StateDegraded State = "degraded"
+)
+
+// Attribution names the dominant cause class when degraded. Two members:
+// unknown is the complement of host, and there is deliberately no internal
+// member — adding one is the answer ENG-3323 exists to measure.
+type Attribution string
+
+const (
+	AttributionUnknown Attribution = "unknown"
+	AttributionHost    Attribution = "host"
+)
+
+// CauseKind enumerates the reason classes that can degrade CPU health.
+type CauseKind string
+
+const (
+	CauseKindSaturation     CauseKind = "saturation"
+	CauseKindThrottling     CauseKind = "throttling"
+	CauseKindPressure       CauseKind = "pressure"
+	CauseKindSteal          CauseKind = "steal"
+	CauseKindHostContention CauseKind = "host-contention"
+)
+
+// Unit is the unit a cause's value is denominated in, copied from the mark pair
+// that judged it so the message layer can render "cores" vs "ratio".
+type Unit string
+
+// cpuReserveCores is the no-limit headroom reserve: one core set aside for
+// Redpanda. It is Redpanda's default maxCores (--smp), not a calibration guess
+// (DECISIONS.md D-08); Decide stamps it onto Signals.ReserveCores so the
+// message reads the same number the verdict subtracted.
+const cpuReserveCores = 1.0
+
+// Cause is one entry in a degraded verdict, ordered by diagnosis.Rank.
+type Cause struct {
+	Kind  CauseKind
+	Value float64
+	Unit  Unit
+}
+
+// Verdict is what Decide returns: the state, the attribution of the dominant
+// cause, and the ranked cause list. The message is NOT a field on it — the
+// message layer composes it from Verdict and Signals.
+type Verdict struct {
+	State       State
+	Attribution Attribution
+	Causes      []Cause
+}
+
+// saturationArm identifies which instrument of the saturation family a folded
+// cause came from. The latch is per signal, so at most one of host-full,
+// no-host-stats and the limit arm survives the fold; attribution differs by
+// arm, so Decide keeps the identity beside the survivor.
+type saturationArm int
+
+const (
+	noSaturationArm saturationArm = iota
+	hostFullArm
+	noHostStatsArm
+	limitArm
+)
